fix(api): Split Game into GameApi and GameApiExtended interfaces, which are implemented by an object
